Created source.lua
This is a script for the game named "Adopt and Raise". The description for this script is below!

This script allows you to get unlimited money by the click of the excecute button on your execute, there is a button for the choice to exit out of the autofarm!
This script is only for the game of the id of "2396588044".

https://www.roblox.com/games/2396588044/Adopt-and-Raise

If you do not do this game then a notification will prompt you asking to teleport you to the game!